- implement user authorization

+++ app/controllers/access_tokens_controller.rb
@@ -1,6 +1,7 @@
 class AccessTokensController < ApplicationController
 
   def create
+    skip_authorization
     user = User.find_by!(email: login_params[:email])
 
     if user.authenticate(login_params[:password])
@@ -28,6 +29,7 @@ class AccessTokensController < ApplicationController
       return
     end
 
+    authorize(access_token)
     access_token.destroy
     render status: :no_content
   end
